requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
google-generativeai==0.3.2
supabase==2.3.0
python-dotenv==1.0.1
streamlit==1.31.1
requests==2.31.0
httpx==0.26.0
websockets==11.0.3
EOL